Refactor LostApiClient connect & permission logic into LostApiClientActivity
