+++ lost-sample/src/main/java/com/example/lost/GeofencingApiActivity.java
@@ -3,34 +3,20 @@ package com.example.lost;
 import com.mapzen.android.lost.api.Geofence;
 import com.mapzen.android.lost.api.GeofencingRequest;
 import com.mapzen.android.lost.api.LocationServices;
-import com.mapzen.android.lost.api.LostApiClient;
 
 import android.app.PendingIntent;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.v4.app.ActivityCompat;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.app.AppCompatActivity;
-import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
-import android.widget.Toast;
 
-import static android.Manifest.permission.ACCESS_FINE_LOCATION;
-import static android.content.pm.PackageManager.PERMISSION_GRANTED;
-import static android.widget.Toast.LENGTH_SHORT;
 import static com.mapzen.android.lost.api.Geofence.NEVER_EXPIRE;
 
 /**
  * Geofencing demo
  */
-public class GeofencingApiActivity extends AppCompatActivity
-    implements LostApiClient.ConnectionCallbacks {
-  private static final String TAG = GeofencingApiActivity.class.getSimpleName();
-
-  private static LostApiClient client;
+public class GeofencingApiActivity extends LostApiClientActivity {
 
   private TextView inputRequestId;
   private TextView inputLatitude;
@@ -40,9 +26,7 @@ public class GeofencingApiActivity extends AppCompatActivity
   @Override protected void onCreate(Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);
     setContentView(R.layout.activity_geofencing);
-    if (client == null) {
-      client = new LostApiClient.Builder(this).addConnectionCallbacks(this).build();
-    }
+
     initConnectButton();
     initDisconnectButton();
     initInputFields();
@@ -87,46 +71,6 @@ public class GeofencingApiActivity extends AppCompatActivity
         }
       });
     }
-  }
-
-  @Override public void onConnected() {
-    Toast.makeText(this, "LOST client connected", LENGTH_SHORT).show();
-  }
-
-  @Override public void onConnectionSuspended() {
-    Toast.makeText(this, "LOST client suspended", LENGTH_SHORT).show();
-  }
-
-  private void connect() {
-    Log.d(TAG, "Connecting...");
-    if (!isFineLocationPermissionGranted()) {
-      requestFineLocationPermission();
-    } else {
-      client.connect();
-    }
-  }
-
-  private boolean isFineLocationPermissionGranted() {
-    return ContextCompat.checkSelfPermission(this, ACCESS_FINE_LOCATION) == PERMISSION_GRANTED;
-  }
-
-  private void requestFineLocationPermission() {
-    ActivityCompat.requestPermissions(this, new String[] { ACCESS_FINE_LOCATION }, 101);
-  }
-
-  @Override public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions,
-      @NonNull int[] grantResults) {
-    if (grantResults[0] == PERMISSION_GRANTED) {
-      client.connect();
-    } else {
-      finish();
-    }
-  }
-
-  private void disconnect() {
-    Log.d(TAG, "Disconnecting...");
-    client.disconnect();
-    Toast.makeText(this, "LOST client disconnected", LENGTH_SHORT).show();
   }
 
   private void createGeofence() {
